(#1016) rename tests, move to io package
