+++ src/test/java/org/cactoos/io/OutputNoNullsTest.java
@@ -21,10 +21,8 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package org.cactoos;
+package org.cactoos.io;
 
-import org.cactoos.io.DeadOutput;
-import org.cactoos.io.OutputNoNulls;
 import org.junit.Test;
 
 /**
@@ -32,7 +30,7 @@ import org.junit.Test;
  * @since 0.10
  * @checkstyle JavadocMethodCheck (500 lines)
  */
-public final class OutputTest {
+public final class OutputNoNullsTest {
 
     @Test(expected = Exception.class)
     public void failForNullOutput() throws Exception {
